Remove legacy push_stack endpoint
