+++ crates/but-server/src/lib.rs
@@ -793,7 +793,6 @@ pub async fn run(config: Config) -> anyhow::Result<()> {
             "/update_branch_pr_number",
             but_post(legacy::stack::update_branch_pr_number_cmd),
         )
-        .route("/push_stack", but_post(legacy::stack::push_stack_cmd))
         // Undo/Snapshot commands
         .route(
             "/list_snapshots",
