+++ crates/gitbutler-branch-actions/Cargo.toml
@@ -21,7 +21,6 @@ export-schema = [
 [dependencies]
 but-error.workspace = true
 but-serde = { workspace = true, features = ["legacy"] }
-but-gerrit.workspace = true
 but-workspace = { workspace = true, features = ["legacy"] }
 but-core = { workspace = true, features = ["legacy"] }
 but-graph = { workspace = true, features = ["legacy"] }
@@ -42,12 +41,11 @@ gitbutler-project.workspace = true
 tracing.workspace = true
 anyhow.workspace = true
 git2.workspace = true
-gix = { workspace = true, features = ["blob-diff", "revision", "merge"] }
+gix = { workspace = true, features = ["blob-diff", "revision"] }
 serde.workspace = true
 bstr.workspace = true
 url = { version = "2.5.4", features = ["serde"] }
 md5.workspace = true
-itertools.workspace = true
 schemars = { workspace = true, optional = true }
 but-schemars = { workspace = true, optional = true }
 
@@ -63,6 +61,7 @@ but-graph.workspace = true
 but-testsupport.workspace = true
 gitbutler-workspace.workspace = true
 gix = { workspace = true, features = [] }
+itertools.workspace = true
 glob = "0.3.3"
 tempfile.workspace = true
 ctor = "0.6.3"
