+++ crates/gitbutler-branch-actions/src/lib.rs
@@ -8,8 +8,6 @@ mod actions;
 // This is our API
 pub use actions::{push_base_branch, set_base_branch, set_target_push_remote};
 
-mod r#virtual;
-
 mod branch_manager;
 pub use branch_manager::BranchManagerExt;
